requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
plotly>=5.18.0
xlsxwriter>=3.1.0
